Change how Dataset values are accessed and processed
Instead of optionally storing the Dataset value as part of the metadata (causing memory issues with large/large number of Datasets), store a reference to the `h5.Dataset` object, from which the value be accessed by the clients that need it.

This required changing where the `h5.File` objects are opened/closed.